Major changes, not quite working all the way yet. moved page table switching to context switch, which avoids it when setting up the new task in spawn_userspace(). added privileged stack in TSS, which allows interrupts to occur while in userspace. Current problem is that context switching away from a userspace task to a kernel thread does not work, there is some page fault error.
I think the best way to resolve this would be to context switch to and from kernel threads only, and to create a "userspace_wrapper" function, much like the kthread_wrapper, that all userspace tasks jump to in their kernel stack.
Then, within that userspace_wrapper function running on the kernel stack, set up everything and jump to userspace. That way, when a context switch occurs, it's just as usual as if it were between kernel threads.