Restore `.sym` parsing
The restored code is largely a duplicate of the code in the
`serialize_nano_core` tool, but it's not quite the same, as the tool
generates `SerializedSection`s whereas `parse_nano_core` generates
`LoadedSection`s. Theoretically we could call the code from
`serialize_nano_core`, generate a `SerializedCrate`, and then load the
crate, but that would add quite a bit of overhead.

Signed-off-by: Klim Tsoutsman <klim@tsoutsman.com>