Implement `raw_mutex` and `mutex`
The current implementation isn't particularly performant, but it is
correct. Improving the performance of the mutex is something we could do
in a later PR.

Unlike `mutex_sleep`, `raw_mutex` lock operations panic rather than
return an error if `task::get_my_current_task()` fails. This conforms to
the `std::sys` mutex API where operations mustn't fail.

`mutex::Mutex` is a copy-paste of `std::sync::Mutex` except it doesn't
keep track of poisoning. As a result, it's lock and unlock operations
are also guaranteed to suceed.

I haven't removed `mutex_sleep` yet because it also implements `RwLock`
which I'll add in a later PR.

Signed-off-by: Klim Tsoutsman <klim@tsoutsman.com>