Implement preemption disabling
This commit is the first step towards closing #578.

The implementation uses a counter over a boolean to keep track of
recursive preemption disabling.

The test application isn't great, but I can't think of a more ergonomic
or thorough way of testing the change.

As a next step (and after #576 lands), we can create a new mutex similar
to `spin::Mutex` and `irq_safety::MutexIrqSafe` that disables preemption
when locked.

Signed-off-by: Klim Tsoutsman <klim@tsoutsman.com>