Separate cargo flags from cargo features.
This commit introduces a new `Makefile` variable called `FEATURES`,
which is passed directly into Theseus's `cargo build` invocation.
It has a default value of `--workspace`, just as `CARGOFLAGS` used to.

This is necessary because `CARGOFLAGS` is reused for correctly
configuring out-of-tree builds against a prebuilt instance of Theseus,
so we cannot overload it with cargo features.

For more info, see the crate-level documentation for `theseus_features`.

At this point, `CARGOFLAGS` is only really used to specify `--release`,
while `FEATURES` is used for controlling which features/crates
are included in a build of Theseus.