Invoke post-context switch action in task_wrapper
* After switching to a new task for the first time, the remainder of the
`task_switch` (and `schedule()`) routines will not execute. To address
this, we define a post-context switch action that is now also invoked
from the task_wrapper to ensure that task states are dropped in a timely
fashion even when switching to a brand new task for the first time.

* We properly clean up bootstrap tasks with a dedicated cleanup task.

* This introduces one small zombie task that we do need to remove in the
future, but that's far better than N tasks that may or may not be
cleaned up quickly based on random task switch ordering, where N is the
number of CPU cores.

* Improve various parts of task-related documentation.