Introduce and clarify post-context switch actions (#600)
* After switching to a new task for the first time, the remainder of the
`task_switch` (and `schedule()`) routines will not execute. To address
this, we define a post-context switch action that is now also invoked
from the task_wrapper to ensure that task states are dropped in a timely
fashion even when switching to a brand new task for the first time.

* We properly clean up bootstrap tasks with a dedicated cleanup task.

* The dedicated cleanup task introduces one small zombie task that
we do need to remove in the future, but that's far better than N tasks
that may or may not be cleaned up quickly based on random
task switch ordering, where N is the number of CPU cores.

* Ensures runstates and a CPU's current task are all set atomically
during a task switch operation.

* Clarify final initialization steps needed for each CPU.

* Improve various parts of task-related documentation. 3f9f2e0c5d1bbdfb80f3af47c0167a26979d9659