Switch to storing run queues in boxed array
Removes the dependency on `atomic_linked_list` which we're trying to
phase out. However, the change means we must to initialise the runqueues
all at once. We do this after initialising the APs and so we can't add
the APs' bootstrap tasks to the runqueue as it doesn't yet exist.

I'm not sure bootstrap tasks are strictly necessary any more, but I'll
leave that for another PR.

Signed-off-by: Klim Tsoutsman <klim@tsoutsman.com>