Add `BootInformation::additional_reserved_memory_regions`
This is the first step in supporting Theseus loaded at an aribtrary
address.

The changes to `frame_allocator` are necessary for the frame allocator
to correctly initialise when a reserved memory region overlaps with more
than one reserved memory region that came before it. Reshuffling the
order that the bootloader info and modules memory regions are added to
the reserved memory regions revealed the problem.

The PR also removes `bios` as the default feature for `nano_core` as it
makes it impossible to compile for UEFI, as discussed on Discord.

Signed-off-by: Klim Tsoutsman <klim@tsoutsman.com>