Fix an issue where a page fault occurs upon a very early panic when `mirror_log_to_vga` is enabled.
This stemmed from the TLS variables that hold the current task and
current task ID being accessed (for fault recovery purposes) upon a
panic, but those TLS data sections aren't generally available until the
tasking subsystem is initialized.

Now, we initialize the TLS data image (using `FsBase` on x86_64`) as
soon as we possibly can, which is right after the `nano_core` is parsed
and first initialized.
We use a dummy/empty TLS data image for this, which ensures that the TLS
register points to a real initialized TLS data image, avoiding weird
problems like an unexpected page fault.

Note that the prior behavior wasn't a real problem or had any safety
implications, it's just caused an annoying issue where you couldn't get
a clean error message when an early error happened on real hardware.