things seem to work.
Only issue is that deallocation isn't behaving as expected?
for ex, an app crate's text sections should be freed after the app is
finished executing and its crate has been unloaded, but that doesn't
seem to be happening.
Running `hello` twice in a row should use the same .text pages for both,
but the second invocation is not re-using the .text pages from the
first invocation.