Rename `scheduler_priority` to `scheduler_epoch`
The short of it is that our priority scheduler isn't actually a priority
scheduler.

I initially made a large PR with a bunch of scheduler changes but
realised that because of all the renaming, file history was lost, so
I've split the PR into a bunch of smaller PRs.

After this is merged I'll create a PR to rename `scheduler_realtime` to
`scheduler_priority`.

Signed-off-by: Klimenty Tsoutsman <klim@tsoutsman.com>