Add `graphic` and `terminal` options to Makefile
`graphic=no` adds the `-nographic` flag to the QEMU invocation.
Supporting this required a few changes to `captain`.

`terminal=(name of terminal)` e.g. `terminal=alacritty` will launch a
new terminal window connected to `SERIAL2`. We do this by testing the
next free TTY with `tools/free_tty` and assuming QEMU will use the same
TTY. This may seem jank, but I think it's the only feasible solution.

I've tested this on MacOS (side note: macOS makes it very hard to launch
a new terminal window from the terminal) so if someone could test this
on Linux as well that would be good. Also, I wasn't able to build for
`aarch64` so the changes to `captain` need to be tested to ensure builds
for `aarch64` didn't break.

Signed-off-by: Klimenty Tsoutsman <klim@tsoutsman.com>