Use priority inheritance for blocking mutex (#985)
Priority inheritance is now used to temporarily boost
the priority of a given task while that task holds a
blocking Mutex lock that another task is waiting on.

In other words, if Task A has a higher priority than Task B,
but Task B holds a mutex that Task A is waiting to acquire,
then we temporarily increase the priority of Task B
to allow it to finish its work inside that critical section
and subsequently release the lock such that Task A
can acquire it.

In this way, we mitigate the effects of priority inversion.
A guard type is used to ensure that Task B's priority
is reset to its original level after it Task A has acquired
the lock that Task B released. 399abef948c553def2948ddc45d78c626ae63257