Remove `Deref` chain between `AllocatedPages` and `PageRange`.
This is problematic for a huge pages implementation,
since the underlying range of pages for a particular
`AllocatedPages` (and `MappedPages`) will not always be
the same single `PageRange` type -- it could be a range
of 4KiB normal pages or 2MiB / 1GiB huge pages.

Instead, `AllocatedPages` now exposes a few wrapper functions
that allow access to the commonly-used functions from `PageRange`.

MappedPages now derefs to `AllocatedPages`, not directly to `PageRange`.