Save `rflags` during context switch
There are no guarantees about the value of the interrupt flag when
context switching. If the context switch is voluntary, i.e. a thread
called `schedule`, interrupts will most likely be enabled, whereas
if a thread is preempted, interrupts will be disabled. But this means
that if a preempted thread A switches to a thread B that voluntarily
yielded, thread B will return from the call to `schedule` with
interrupts disabled.

The AArch64 code also needs to be modified but I'll leave that to
@NathanRoyer.

Signed-off-by: Klimenty Tsoutsman <klim@tsoutsman.com>