Rename `async_channel` to `sync_channel` (#1078)
The previous `async_channel` was asynchronous in the sense
that it had an internal buffer that allowed a sender and receiver
to not have to synchronously rendezvous in order to exchange
a message/data. However, it was not asynchronous in the sense
of language-level async/await, so the naming would be confusing. 

We rename that channel implementation to `sync_channel`, 
which is more accurate in the sense that Senders and Receivers 
interact with it synchronously (not using async/await), which means
the current task may be blocked when waiting on a send or receive 
operation. 

This is in preparation for an upcoming PR that will introduce a channel
that supports language-level async/await syntax. 

Signed-off-by: Klimenty Tsoutsman <klim@tsoutsman.com>