Lazily remove exited tasks from run queue in scheduler
Avoids locking all the schedulers on task exit.

This PR also technically changes the round robin scheduler algorithm, as
blocked tasks are now moved to the end of the run queue. Previously, the
blocked tasks would be kept in place, aside from the task at the front of
the queue which would be switched with the next runnable task using
`swap_remove_front`.

Signed-off-by: Klim Tsoutsman <klim@tsoutsman.com>